mcp/python-openai-mcp/requirements.txt
openai>=2.4.0
python-dotenv>=1.1.1

